Map connections
This class creates a linked list for all of the nodes in the map.

you can you it by sending in the;  ID of node that has the connections
coming from, ID of the first node you are connecting to, ID of the
second node, ID of the third connection node, and lastly the connection
of the fourth node.

It would be most efficent to initialise using a text file.